+++ src/main/kotlin/org/patternfly/Settings.kt
@@ -3,6 +3,7 @@ package org.patternfly
 // TODO Document me
 public object Settings {
 
+    public const val DRAWER_INCREMENT: Int = 5
     public const val NOTIFICATION_TIMEOUT: Int = 8000 // ms
     public const val UI_TIMEOUT: Long = 333L // ms
 }